Implemented functions to get and update the 'status' of the user's local CDragon instance
The status is based on the last time their 'rcp-be-lol-game-data' plugin was updated.

When the user fetches the latest data, the status of their application goes from 'uninitialized' to 'up-to-date'. If the plugin is updated at https://raw.communitydragon.org/json/latest/plugins then their cached data is now stale. The 'update' function can be called to fetch the data and update the cache.